backend-fastapi/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
llama-cpp-python
sse-starlette
numpy<2.0
sentence-transformers
torch
transformers
